refactor: extract inline tests from remaining modules
Extract #[cfg(test)] mod tests blocks from 11 source files into
separate test files following the existing pattern:

- copilot_task_submit/orchestration.rs → tests_orchestration.rs
- copilot_task_submit/transcript.rs → tests_transcript.rs
- improvements/parsing.rs → tests_parsing.rs
- improvements/promotion.rs → tests_promotion.rs
- ooda_actions/goal_session.rs → tests_goal_session.rs
- ooda_loop/observe.rs → tests_observe.rs
- operator_cli/mod.rs → tests_mod.rs
- operator_commands/dispatch.rs → tests_dispatch.rs
- operator_commands_meeting/live_context.rs → tests_live_context.rs
- self_improve/cycle.rs → tests_cycle.rs
- bootstrap/config.rs → tests_config.rs

Private items needed by tests promoted to pub(super):
- build_evidence_records, OPERATOR_CLI_HELP
- skip_record_separators, skip_spaces, looks_like_field_start
- collect_pending_improvements, scan_unprocessed_handoffs_in
- next_required, next_optional_path, reject_extra_args
- decide, find_weak_dimensions
