+++ src/improvements/mod.rs
@@ -3,6 +3,11 @@ mod persisted;
 mod promotion;
 mod types;
 
+#[cfg(test)]
+mod tests_parsing;
+#[cfg(test)]
+mod tests_promotion;
+
 // Re-export all public items so `crate::improvements::X` still works.
 pub use promotion::render_review_context_directives;
 pub use types::{
